fix: Do not flicker Docusaurus logo on load
Previously, the Docusaurus logo would appear twice for a short while during
load. This was due to the `data-theme` attribute we set on the body not
matching the `data-theme` attribute of the document. This happened because
during hydration, the `useIsBrowser()` hook always returns `false`.

We fix this by no longer using any React info that is affected by the
hydration quirk, and instead update the attribute using pure JS DOM APIs in
a useEffect call.

Fixes the issue reported in https://github.com/facebook/docusaurus/issues/9527